refactor(app): move queries folder inside lib folder
