+++ packages/app/src/App.tsx
@@ -15,8 +15,8 @@ import PlayerHeadBakingDialog from './components/dialog/PlayerHeadBakingDialog'
 import PromptDialog from './components/dialog/PromptDialog'
 import SettingsDialog from './components/dialog/SettingsDialog'
 import WelcomeDialog from './components/dialog/WelcomeDialog'
+import { queryClient } from './lib/query.ts'
 import AutosaveService from './lib/services/autosave.service.ts'
-import { queryClient } from './query.ts'
 import { useDialogStore } from './stores/dialogStore'
 import { useEditorStore } from './stores/editorStore'
 import { useProjectStore } from './stores/projectStore.ts'
