+++ packages/app/src/components/dialog/ItemDisplaySelectDialog.tsx
@@ -4,7 +4,7 @@ import { type FC, useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useShallow } from 'zustand/shallow'
 
-import { getItemListQueryFn } from '@/queries/getItemList'
+import { getItemListQueryFn } from '@/lib/queries/getItemList'
 import { useDialogStore } from '@/stores/dialogStore'
 import { useDisplayEntityStore } from '@/stores/displayEntityStore'
 import { useProjectStore } from '@/stores/projectStore'
