+++ packages/app/src/lib/queries/getBlockList.ts
@@ -1,7 +1,7 @@
 import type { QueryFunctionContext } from '@tanstack/react-query'
 
 import { CDNBaseUrl } from '@/constants'
-import { queryClient } from '@/query'
+import { queryClient } from '@/lib/query'
 import type { CDNBlocksListResponse } from '@/types'
 
 export async function getBlockListQueryFn({
